% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantisdiagnostics.r
\docType{package}
\name{atlantisdiagnostics}
\alias{atlantisdiagnostics}
\title{atlantisdiagnostics: Tools for calibrating Atlantis ecosystem models}
\description{
Tools to aid in diagnosing issues in a parameterized Atlantis Ecosystem model
}
\details{
\itemize{
\item Determine functional group persistence over time
\item Determine functional group stability over time
\item Determine functional group biomass (relative to historic data) lie within reasonable bounds
\item Determine distribution of biomass over age classes
\item Determine maximum size (length and weight) of species relative to observations

}
}
\section{References}{


To learn more about using \code{atlantisdiagnostics}, start with the vignette: \code{browseVignettes(package="atlantisdiagnostics")} or click the index below
}

