% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mef.R
\name{calc_mef}
\alias{calc_mef}
\title{Calculate Modeling Efficiency (MEF)}
\usage{
calc_mef(observed, expected)
}
\arguments{
\item{observed}{observed values (data)}

\item{expected}{expected values (model)}
}
\value{
list of two elements
\item{mef}{model efficiency value: Range from -Inf to 1}
\item{n}{number of observations used to calculate mef}
}
\description{
Internal function to calculate modeling efficiency
}
