% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_params.R
\name{check_params}
\alias{check_params}
\title{Check parameters exist}
\usage{
check_params(prmDir, fgs)
}
\arguments{
\item{prmDir}{Character vector. Path to biology parameter file}

\item{fgs}{Data frame. Contents of the functional group file. \code{atlantisom::load_fgs}}
}
\value{
Returns a data frame of species which do not meet defined persistence criteria.

\item{a}{The common name of the species/functional group}
}
\description{
Inspects biology parameter file to ensure all parameter values have been entered. Missing values will not
cause the model to crash but will produce unexpected results. Explain ...
}
\examples{
\dontrun{
}
}
