% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_species_codes.R
\name{check_species_codes}
\alias{check_species_codes}
\title{Check for valid species codes}
\usage{
check_species_codes(modelBiomass, speciesCodes)
}
\arguments{
\item{modelBiomass}{Data frame. Output from \code{atlantisom::load_bioind}}

\item{speciesCodes}{Character vector. Vector of species Codes to verify against complete list}
}
\value{
Vector of valid Codes
}
\description{
Uses complete list of atlantis codes to ensure codes passed to test are part of the list
}
