% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_cohortBiomass.r
\name{diag_cohortBiomass}
\alias{diag_cohortBiomass}
\title{Test for distribution of biomass over age class}
\usage{
diag_cohortBiomass(mortFile, agebiomind, neusPriority)
}
\arguments{
\item{mortFile}{A character string. Path to location of Mort.txt file.}

\item{agebiomind}{A character string. Path to location of AgeBiomIndx file.}

\item{neusPriority}{A character string. Path to location of Species Priorities file.}
}
\value{
\item{Code}{Atlantis species code}
\item{Status}{Boolean indicating whether species passes test}
\item{Max_Cohort}{Age class with the highes biomass}
\item{Stability}{}
\item{Priority}{Scale defining priority of species in the model}
\item{Fishing}{Scale defining level of fishing, 1 is highest}
}
\description{
Determine which age classes contain the most biomass and compare to neighboring
age classes. A species fails the test is all biomass is contained in either
the smallest or largest age class
}
