% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_combine.R
\name{diag_combine}
\alias{diag_combine}
\title{Combine all diagnostic output}
\usage{
diag_combine(persistence, stability, reasonability)
}
\arguments{
\item{persistence}{Tibble. Output from \code{diag_persistence}}

\item{stability}{Tibble. Output from \code{diag_stability}}

\item{reasonability}{Tibble. Output from \code{diag_reasonability}}
}
\value{
A data frame
\item{species}{The common name of the species/functional group as described in Atlantis input file}
\item{code}{Atlantis Code for species/functional group}
\item{persistence}{Indicating pass or failure of persistence test}
\item{proportionInitBio}{Metric for persistence. The proportion of initial biomass at \code{tminimumBiomass}}
\item{stability}{Indicating pass or failure of stability test}
\item{relChange}{Metric for stability. Rate of increase relative to \code{t1Biomass}(\code{mtperyear}/\code{t1Biomass}). see \code{\link{diag_stability}} }
\item{reasonability}{Indicating pass or failure of reasonability test}
}
\description{
Using the output from all tests (persistence, stability, reasonability), they are combined
into a single data frame
}
