% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_fleet_catch.R
\name{diag_fleet_catch}
\alias{diag_fleet_catch}
\title{Test fleet catch}
\usage{
diag_fleet_catch(
  fgs,
  fishery.prm,
  catch.file,
  bgm,
  catch.ref,
  speciesCodes = NULL,
  nYrs = 20,
  min.dist = 100,
  relChangeThreshold = 0.01
)
}
\arguments{
\item{fgs}{A character string. Path to location of functional groups file.}

\item{fishery.prm}{A character string. Path to the fishery definitions file.}

\item{catch.file}{A character string. Path to catch nc file}

\item{bgm}{A character string. Path to the bgm file.}

\item{catch.ref}{A data.frame containing reference catch by fleet data (species|fleet|polygon|ref.value)}

\item{speciesCodes}{Character vector. A vector of Atlantis species codes in which to test for stability.}

\item{nYrs}{Numeric scalar. Number of years from the end of the time series that stability must occur.}

\item{min.dist}{Numeric Scalar. Maximum distance between model and reference center of gravity}

\item{relChangeThreshold}{Numeric Scalar. Maximum magnitude of relative change of slope (Default = 0.01)}
}
\value{
Returns a data frame of all species and how they measure up against the catch by fleet criteria
\item{species}{The common name of the species/functional group}
\item{fleet}{The fleet name from the fisheries file}
\item{catch.model}{Double. The mean catch in mT over the last nYrs of the model }
\item{catch.ref}{Double. The reference catch (mT)}
\item{cog.x.model}{Double. The mean x-coordinate center of gravity from the model}
\item{cog.x.ref}{Double. The mean x-coordinate center of gravity from the reference}
\item{cog.y.model}{Double. The mean y-coordinate center of gravity from the model}
\item{cog.y.ref}{Double. The mean y-coordinate center of gravity from the reference}
\item{dist}{Double. The mean distance between the center of gravity between the model and reference}
\item{catch.magnitude}{Logical. Is the relative difference between model and refernece catch within +/- relChangeThreshold?}
\item{catch.dist}{Logical. Is the distance between center of gravity betwene model and reference less than min.dist?}
}
\description{
\code{diag_fleet_catch} determines whether the simulated fleets catcht the right magnitude and spatial distribution of catch
over the last n years of a run.
}
\seealso{
Other diags: 
\code{\link{diag_persistence}()},
\code{\link{diag_reasonability}()},
\code{\link{diag_stability}()}
}
\concept{diags}
