% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_init.R
\name{diag_init}
\alias{diag_init}
\title{Initialize Atlantis diagnostics}
\usage{
diag_init(config_file)
}
\arguments{
\item{config_file}{file specifying google directory with atlantis model outputs and a local
directory for download, and identifying key input files. See example config_file below.}
}
\value{
Returns a diaglist list object containing dataframes and lists:
\itemize{
\item{fgs, dataframe of species group characteristics}
\item{fgs.names, vector of species names}
\item{atBtxt, dataframe output of \code{atlantisom::load_bioind}}
\item{atCtxt, dataframe output of \code{atlantisom::load_catch}}
\item{runpar, list of run parameters output of \code{atlantisom::load_runprm}}
\item{noutsteps, integer number of output steps in the model run}
\item{stepperyr, integer number of putput steps per year}
\item{timeall, vector of all model timesteps}
\item{fstepperyr, integer fishery number output steps per year}
\item{boxpars, dataframe of spatial parameters}
\item{boxall, vector of all model box numbers}
\item{YOY, dataframe young of year output}
\item{biol, list of biological parameters output of \code{atlantisom::load_biolprm}}
}
}
\description{
\code{diag_init} pulls necessary Atlantis files from google drive and loads
them into R objects for input into each diagnostic function.
}
\examples{
\dontrun{
# Example config file DiagnosticConfig.R:
#------------------------------------
# CHECK ALL FILENAMES FOR YOUR MODEL

# where are files on the atlantis google drive?
g.name <- "Testing/OutForSarah"

# which local directory should they write to?
d.name <- here::here("diagnostics", "temp")

# input file names
functional.groups.file <- "neus_groups.csv"
biomass.pools.file <- "neus_init.nc"
box.file <- "neus_tmerc_RM2.bgm"
initial.conditions.file <- "neus_init.nc"
fisheries.file <- "neus_fisheries.csv"
biol.prm.file <- "at_biology.prm"

# define the hindcast period--these are survey years
hindcast <- c(1980:2010)
#------------------------------------

#Usage
testdiag <- diag_init("DiagnosticConfig.R")
}


}
\author{
Sarah Gaichas
}
