% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_maxsize.R
\name{diag_maxsize}
\alias{diag_maxsize}
\title{Test for max fish size}
\usage{
diag_maxsize(
  nc,
  bgm,
  init,
  fgs,
  prm_run,
  prm_biol,
  speciesStats,
  speciesCodes = NULL
)
}
\arguments{
\item{nc}{A character string. Path to location of model run output files.}

\item{bgm}{A character string. Prefix added to all atlantis output files.}

\item{init}{A character string. Prefix added to all atlantis output files.}

\item{fgs}{A character string. Prefix added to all atlantis output files.}

\item{prm_run}{A character string. Prefix added to all atlantis output files.}

\item{prm_biol}{A character string. Prefix added to all atlantis output files.#'}

\item{speciesStats}{Data frame.}

\item{speciesCodes}{Character vector. A vector of Atlantis species codes in which to test for persistence.
(Default = NULL, uses all species)}
}
\value{
Returns a data frame of species which do not meet defined max size criterion.

\item{species}{The common name of the species/functional group}
\item{code}{Atlantis Code for species/functional group}
}
\description{
Calculate average size of fish through time and check to see if any fish exceed the
user supplied observed maximum size.
}
