% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_persist.R
\name{diag_persist}
\alias{diag_persist}
\title{Test for all functional group persistence}
\usage{
diag_persist(atBtxt, fgs.names, plot = TRUE)
}
\arguments{
\item{atBtxt}{A dataframe of total biomass of all groups over time, read in from
Atlantis ...BioInd.txt output using \code{atlantisom::load_bioind}.}

\item{fgs.names}{A vector of species names}

\item{plot}{A logical value specifying if the function should generate plots or
not. The default is \code{TRUE}.}
}
\value{
Returns a dataframe of crashed species with columns "Failed" of species
names and "nYrs0B" with number of years having 0 biomass. Also plots all species
trajectories with red lines for crashed years unless plot=FALSE.
}
\description{
\code{diag_persist} Averages over a year to see if mean B is 0,
assuming that initial (Time = 0) biomass does not count. Output is a list of
species that have crashed and an optional set of plots where 0 biomass is red.
}
\details{
This should be run on an unfished, unperturbed run. Fishing or
perturbations may legitimately drive groups extinct. However, for our
historical period, we don’t expect anything to go extinct that is in the
system now, so this test includes historical fishing, physics, and
phytoplankton drivers.
}
\examples{
\dontrun{
# Usage
diag_persist(testdiag$atBtxt, testdiag$fgs.names)
}

}
\author{
Sarah Gaichas
}
