% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_persistence.R
\name{diag_persistence}
\alias{diag_persistence}
\title{Test for functional group persistence}
\usage{
diag_persistence(
  modelBiomass,
  speciesCodes = NULL,
  nYrs = NULL,
  floor = 0.1,
  display = NULL,
  tol = 1e-06,
  plot = F
)
}
\arguments{
\item{modelBiomass}{A data frame. Total biomass of all groups over time, read in from
Atlantis ...BioInd.txt output using \code{atlantisom::load_bioind}.}

\item{speciesCodes}{Character vector. A vector of Atlantis species codes in which to test for persistence.
(Default = NULL, uses all species found in  \code{modelBiomass})}

\item{nYrs}{Numeric scalar. Number of years from the end of the time series that persistence must occur.
(Default = NULL, persistence must occur throughout entire time series)}

\item{floor}{Numeric scalar. Proportion of initial biomass for which for which all species
are measured against. (Default = 0.1, all species need to be above 10\% of initial biomass). Range should be 0-1}

\item{display}{Boolean. Flag to indicate whether to return only species that pass the test, fail the test, or all species (Default = NULL)}

\item{tol}{Numeric scalar. Tolerance level to add to biomass floor (Default = 1E-6)}

\item{plot}{A logical value specifying if the function should generate plots or
not. (Default = F).}
}
\value{
Returns a data frame of species which do not meet defined persistence criteria.

\item{species}{The common name of the species/functional group}
\item{code}{Atlantis Code for species/functional group}
\item{initialBiomass}{Starting value of Biomass for species/functional group}
\item{minimumBiomass}{The smallest value of biomass observed in the run}
\item{tminimumBiomass}{The time step in which \code{minimumBiomass} occurred }
\item{proportionInitBio }{The proportion of initial biomass at \code{tminimumBiomass}}
\item{t1}{The first time step persistence was not met}
\item{tn}{The last time step persistence was not met}
\item{nts}{The total number of time steps that persistence was not met}
\item{pass}{Boolean. Indicates if the Atlantis group passed persistence test}
}
\description{
Inspects each time point. If at any point in time a species falls below a predefined floor it
is flagged. The floors are specifies as a proportion of initial biomass. The last n years of the run are
used in the test
}
\examples{
\dontrun{
# Declare paths to files required
biol.file <- "neus_outputBiomIndx.txt"
file_fgs <- "neus_groups.csv"
# use atlantisom to read them in
fgs <- atlantisom::load_fgs(inDir,file_fgs)
modelBiomass <- atlantisom::load_bioind(outDir,biol.file,fgs)

# find all species that do not have biomass > 0 for any time during the run.

diag_persistence(modelBiomass,speciesCodes=NULL, nYrs = NULL, floor = 0)

# only evaluate herring. Require stability over the last 10 years of the run and all values should
# exceed 10\% of initial biomass
diag_persistence(modlBiomass, speciesCodes="HER", nYrs = 10, floor = 0.1)

}
}
