% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_reasonability.R
\name{diag_reasonability}
\alias{diag_reasonability}
\title{Test for all functional groups to persist within reasonable bounds}
\usage{
diag_reasonability(
  modelBiomass,
  initialYr = 1964,
  speciesNames = NULL,
  realBiomass,
  nYrs = NULL,
  bounds = c(1, 1),
  tol = 1e-06,
  plot = F
)
}
\arguments{
\item{modelBiomass}{A data frame. Total biomass of all groups over time, read in from
Atlantis ...BioInd.txt output using \code{atlantisom::load_bioind}.}

\item{initialYr}{Numeric Scalar. Year in which the model run was initiated. (Default = 1964)}

\item{speciesNames}{Character vector. A vector of species names in which to test for reasonableness
(Default = NULL, uses all species found in  \code{biomass}. Results will depend on availability of calculable
bounds based on "data")}

\item{realBiomass}{A data frame. biomass time series (from assessments, stock SMART or otherwise) for species.
\code{realBiomass} should be in long format with at least 3 columns labeled Year, Value, CommonName.}

\item{nYrs}{Numeric scalar. Number of years from the end of the time series for which reasonableness is checked.
(Default = NULL, entire time series is used)}

\item{bounds}{Numeric vector. Size of 1x2 containing the values in which to multiple lower and upper bounds of observed data.
For example (Default = c(1,1)) indicating use of min and max of observed biomass
Proportion of initial biomass for which for which all species
are measured against. (Default = 0, all species need to be non zero). Range should be 0-1}

\item{tol}{Numeric scalar. Tolerance level to add to biomass floor (Default = 1E-6)}

\item{plot}{A logical value specifying if the function should generate plots or
not. (Default = F).}
}
\value{
Returns a data frame of species

\item{species}{The common name of the species/functional group}
\item{Code}{Atlantis Code for species/functional group}
}
\description{
Inspects each time point. If at any point in time a species falls below or rises above predefined
bounds it is flagged. The term "reasonable" is based on survey data when available.
}
\examples{
\dontrun{
}
}
