% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_stability.R
\name{diag_stable}
\alias{diag_stable}
\title{Test for model stability}
\usage{
diag_stable(atBtxt, fgs.names, runpar, plot = TRUE)
}
\arguments{
\item{atBtxt}{A dataframe of total biomass of all groups over time, read in from
Atlantis ...BioInd.txt output using \code{atlantisom::load_bioind}.}

\item{fgs.names}{A vector of species names}

\item{runpar}{A list of Atlantis run parameters read in from the .xml file using
\code{atlantisom::load_runprm}}

\item{plot}{A logical value specifying if the function should generate plots or
not. The default is \code{TRUE}.}
}
\value{
Returns a dataframe of crashed species with columns "Failed" of species
names and "nYrs0B" with number of years having 0 biomass. Also plots all species
trajectories with orange (positive) or purple (negative) trend lines for unstable
groups over the last 20 years unless plot=FALSE.
}
\description{
\code{diag_stable} determines whether the model reaches steady state
for the last ~20 years of an unperturbed, unfished ~100 year run.
}
\details{
This test evaluates the last 20 years of the input run for a significant
slope using the \code{geom_gls} function from the package \code{ecodata} at
\link{https://github.com/NOAA-EDAB/ecodata}. The function \code{diag_stable} also
checks for persistence and flags groups that have gone to 0 biomass--they don't
count as stable. It takes the same inputs as \code{diag_persist} with the addition of
the \code{runpar} object, and defaults to plotting, which can be turned off with
\code{plot=FALSE} in the function call.
}
\examples{
\dontrun{
# Usage
diag_stable(testdiag$atBtxt, testdiag$fgs.names, testdiag$runpar)
}

}
\author{
Sarah Gaichas
}
