% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_cohortBiomass.r
\name{diag_cohortBiomass}
\alias{diag_cohortBiomass}
\title{Test for distribution of biomass over age class}
\usage{
diag_cohortBiomass(
  fgs,
  mortality,
  agebiomind,
  speciesCodes = NULL,
  neusPriority,
  fishedSpecies = 2
)
}
\arguments{
\item{fgs}{A character string. Path to location of functional groups file.}

\item{mortality}{A character string. Path to location of Mort.txt file.}

\item{agebiomind}{A character string. Path to location of AgeBiomIndx file.}

\item{speciesCodes}{Character vector. A vector of Atlantis species codes in which to test for persistence.
(Default = NULL, uses all species with \code{IsTurnedOn=1} in \code{fgs} file)}

\item{neusPriority}{A character string. Path to location of Species Priorities file.}

\item{fishedSpecies}{Numeric scalar. Filter by how much species are fished. Values = 1 (most), 2, 3, 4 (least)
(Default = 2. Only species fished at level <= 2 will be returned)}
}
\value{
\item{code}{Atlantis species code}
\item{status}{Boolean indicating whether species passes test}
\item{maxCohort}{Age class with the highes biomass}
\item{stability}{}
\item{priority}{Scale defining priority of species in the model, High (H), Low(L) }
\item{fishing}{Scale defining level of fishing, 1 is highest}
}
\description{
Determine which age classes contain the most biomass and compare to neighboring
age classes. A species fails the test if all biomass is contained in either
the smallest or largest age class
}
