% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_reasonability.R
\name{diag_reasonability}
\alias{diag_reasonability}
\title{Test for functional group reasonability}
\usage{
diag_reasonability(
  modelBiomass,
  initialYr = 1964,
  speciesNames = NULL,
  realBiomass,
  nYrs = NULL,
  surveyBounds = c(1, 1),
  initBioBounds = 0.2,
  plot = F
)
}
\arguments{
\item{modelBiomass}{A data frame. Total biomass of all groups over time, read in from
Atlantis ...BioInd.txt output using \code{atlantisom::load_bioind}.}

\item{initialYr}{Numeric Scalar. Year in which the model run was initiated. (Default = 1964)}

\item{speciesNames}{Character vector. A vector of species names in which to test for reasonableness
(Default = NULL, uses all species found in  \code{biomass}. Species names must be a subset of the Atlantis species names.}

\item{realBiomass}{A data frame. biomass time series (from assessments, stock SMART or otherwise) for species.
CURRENTLY IMPLEMENTED USING SURVDAT DATA ONLY  \code{realBiomass} should be in long format with column labels (YEAR,variable,value,Code,Species,isFishedSpecies)
Biomass units should be in metric tonnes}

\item{nYrs}{Numeric scalar. Number of years from the end of the time series for which reasonableness is checked.
(Default = NULL, entire time series is used)}

\item{surveyBounds}{Numeric vector. Size of 1x2 containing the values in which to multiple lower and upper bounds of observed data.
For example (Default = c(1,1)) indicating use of min and max of observed biomass}

\item{initBioBounds}{Numeric scalar. Proportion of initial biomass. Allowable amount in which species are allowed to deviate. This is used for groups/species that dont
have surveys}

\item{plot}{A logical value specifying if the function should generate plots or
not. (Default = F). NOT YET IMPLEMENTED}
}
\value{
Returns a data frame of species

\item{species}{The common name of the species/functional group as described in Atlantis input file}
\item{code}{Atlantis Code for species/functional group}
\item{initialBiomass}{Starting value of Biomass for species/functional group. From model output}
\item{minimumBiomass}{The smallest value of biomass observed in the run}
\item{maximumBiomass}{The largest value of biomass observed in the run}
\item{t1}{The first year reasonableness was not met}
\item{tn}{The last year reasonableness was not met}
\item{nts}{The total number of years that reasonableness was not met}
\item{modelSkill}{Measure of the goodness of fit}
\item{nObs}{number of observations used in the goodness of fit}
\item{pass}{Boolean. Indicates if the Atlantis group passed reasonability}
\item{test}{Character. Indicate how \code{pass} was determined. If the Atlantis species is fished and the species is a stand alone species in the model
then survey data is used to assess reasonability. Otherwise deviations from initial biomass is used}
}
\description{
Inspects each time point. If at any point in time a species falls below or rises above predefined
bounds it is flagged. The term "reasonable" is based on survey data when available.
}
\examples{
\dontrun{
}
}
