% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_stability.R
\name{diag_stability}
\alias{diag_stability}
\title{Test for functional group stability}
\usage{
diag_stability(
  modelBiomass,
  initialYr = 1964,
  speciesCodes,
  nYrs = 20,
  relChangeThreshold = 0.01,
  plot = F
)
}
\arguments{
\item{modelBiomass}{A data frame. Total biomass of all groups over time, read in from
Atlantis ...BioInd.txt output using \code{atlantisom::load_bioind}.}

\item{initialYr}{Numeric Scalar. Year in which the model run was initiated. (Default = 1964)}

\item{speciesCodes}{Character vector. A vector of Atlantis species codes in which to test for stability.
(Default = NULL, uses all species found in  \code{modelBiomass})}

\item{nYrs}{Numeric scalar. Number of years from the end of the time series that stability must occur.
(Default = 20 years)}

\item{relChangeThreshold}{Numeric Scalar. Maximum magnitude of relative change of slope (Default = 0.01)}

\item{plot}{Logical. Specifying whether the function should generate plots or not. (Default is F).}
}
\value{
Returns a data frame of all species and how they measure up against the stability criterion
\item{code}{Atlantis Code for species/functional group}
\item{species}{The common name of the species/functional group}
\item{t1Fit}{Value of fitted biomass for the first of year data used in the fit}
\item{mtPerYear}{Double. The value of the slope parameter (year) }
\item{relChange}{Rate of increase relative to \code{t1Fit}(\code{mtPerYear}/\code{t1Fit})}
\item{aveBio}{mean biomass for the last \code{nYrs} years }
\item{pass}{Logical. Does the species/group pass the test for stability}
}
\description{
\code{diag_stability} determines whether the model reaches a steady state
over the last n years of a run. Stability is loosely defined as a species/groups biomass reaching a
stable level (measured as having tolerable trend in a defined time range).
}
\section{Details}{


Formally the following model is fit to the last n years of the run:

\deqn{biomass_t = \mu + \beta.t + \epsilon_t  where \epsilon_t ~ IID N(0,\sigma^2)}

where null hypothesis,  \deqn{H0:\beta=0}

Note: annual biomass is used in fitting. Species with mean annual biomass < 1 metric ton over the last
n years of the run are not considered stable. They are reported to Fail the test and NaNs returned
}

\examples{
\dontrun{
# Declare paths to files required
biol.file <- "neus_outputBiomIndx.txt"
file_fgs <- "neus_groups.csv"
# use atlantisom to read them in
fgs <- atlantisom::load_fgs(inDir,file_fgs)
modelBiomass <- atlantisom::load_bioind(outDir,biol.file,fgs)

# Perform stability test on all species/groups using the last 20 years of the run
diag_stability(modelBiomass, nYrs = 20)

# Only perform test on herring and white hake.
# Require stability over the last 10 years of the run and and use alpha = 0.1 as
tests significance level
diag_stability(modelBiomass, speciesCodes=c("HER","WHK"), nYrs = 10, relChangeThreshold = 0.01)
}
}
\seealso{
Other diags: 
\code{\link{diag_reasonability}()}
}
\concept{diags}
