% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_stability.R
\name{diag_stability}
\alias{diag_stability}
\title{Test for functional group stability}
\usage{
diag_stability(
  fgs,
  biomind,
  initialYr = 1964,
  speciesCodes,
  nYrs = 20,
  relChangeThreshold = 0.01
)
}
\arguments{
\item{fgs}{A character string. Path to location of functional groups file.}

\item{biomind}{A character string. Path to the BiomIndx.txt file.}

\item{initialYr}{Numeric Scalar. Year in which the model run was initiated. (Default = 1964)}

\item{speciesCodes}{Character vector. A vector of Atlantis species codes in which to test for stability.
(Default = NULL, uses all species found in  \code{modelBiomass})}

\item{nYrs}{Numeric scalar. Number of years from the end of the time series that stability must occur.
(Default = 20 years)}

\item{relChangeThreshold}{Numeric Scalar. Maximum magnitude of relative change of slope (Default = 0.01)}
}
\value{
Returns a data frame of all species and how they measure up against the stability criterion
\item{code}{Atlantis Code for species/functional group}
\item{species}{The common name of the species/functional group}
\item{t1Fit}{Value of fitted biomass for the first of year data used in the fit}
\item{mtPerYear}{Double. The value of the slope parameter (year) }
\item{relChange}{Rate of increase relative to \code{t1Fit}(\code{mtPerYear}/\code{t1Fit})}
\item{aveBio}{mean biomass for the last \code{nYrs} years }
\item{pass}{Logical. Does the species/group pass the test for stability}
}
\description{
\code{diag_stability} determines whether the model reaches a steady state
over the last n years of a run. Stability is loosely defined as a species/groups biomass reaching a
stable level (measured as having tolerable trend in a defined time range).
}
\section{Details}{


Formally the following model is fit to the last n years of the run:

\deqn{biomass_t = \mu + \beta.t + \epsilon_t  where \epsilon_t ~ IID N(0,\sigma^2)}

where null hypothesis,  \deqn{H0:\beta=0}

Note: annual biomass is used in fitting. Species with mean annual biomass < 1 metric ton over the last
n years of the run are not considered stable. They are reported to Fail the test and NaNs returned
}

\examples{
\dontrun{
# Declare paths to files required

biomind <- paste("Full path to file","xxxBiomIndx.txt")
fgs <- paste("Full path to file","functioalGgroups.csv")

# Perform stability test on all species/groups using the last 20 years of the run
diag_stability(fgs, biomind, nYrs = 20)

# Only perform test on herring and white hake.
# Require stability over the last 10 years of the run and and use a
# relative change in the slope = 0.01 as the criterion for pass or fail
diag_stability(fgs,biomind, speciesCodes=c("HER","WHK"), nYrs = 10, relChangeThreshold = 0.01)
}
}
\seealso{
Other diags: 
\code{\link{diag_fleet_catch}()},
\code{\link{diag_persistence}()},
\code{\link{diag_reasonability}()}
}
\concept{diags}
